Revert "Create separate files for the modules (#326)" (#327)
This reverts commit b618c29217d34d95b9d85151552bd6d61df25e6f.
