--- app/src/main/java/org/wikiedufoundation/wikiedudashboard/di/persistenceModule.kt
@@ -1,15 +1,0 @@
-package org.wikiedufoundation.wikiedudashboard.di
-
-import org.koin.dsl.module
-import org.wikiedufoundation.wikiedudashboard.data.preferences.SharedPrefs
-
-/**
- * Use the [persistenceModule] to creating shared preference instance
- **/
-val persistenceModule = module {
-
-    /**
-     * Singleton for shared preference
-     **/
-    single { SharedPrefs(get()) }
-}
